+++ CHANGELOG.md
@@ -14,6 +14,8 @@
 - Renamed `WsjcppCore::getListOfDirs` to `wsjcpp::directory_list`
 - Fixed #40: Missing first character in name of child directory
 - Removed definition `WsjcppCore::listOfFiles` because missing implementation
+- Fixed #36: Renamed `WsjcppCore::createUuid` to `wsjcpp::generate_uuid()`
+
 
 ## [v0.2.5] - 2026-05-19 (2026 May 19)